extract sample_entity_extraction
